Allow 0 timestamps
In the past we never allowed 0 timestamps. We relaxed this to allow 0
timestamp application values (usegful for migrated data), but did not
account for these timestamps then ending up in snapshots and
subsequently being loaded by LS, triggering the timestamp check.

Additionally, switch our default timestamp for old entries when using
shadow databases from the year 2000 to one 1 ns after the UNIX epoch
(1970). We use a non-zero value to not trigger the 0 check on older
clients.