Add override_create_flags to dbi_options
Allow overriding the DBI create flags on a per-DBI basis in case we try
to load a snapshot and the target DBI does exist yet.

This can be useful to support loading old formatVersion < 3 snapshot,
since they do no contain the right flags, but it should be avoided
unless you are sure you need it.