proto: patch generated code to reduce memory use
The generated ptotobuf code made copies of all key/value byte slices.
This patches the generated code to directly return the byte slices.

In local testing, this reduced memory allocations when loading snapshots
by about 60%.

Next step is to rewrite the protobuf handling to iterate over the
protobuf data instead of creating the []KV slices, but that requires a
significant rewrite.