Pass lmdb.Env to Syncer, fix flaky sync tests
Explicitly pass the lmdb.Env to the syncer, so that we have more control
over its lifetime and avoid opening and closing it multiple times, which
can cause crashes in the LMDB C code.

Rewrite the sync tests to wait for expected values, instead of relying
on arbitrary sleeps that make the test flaky and slow.

Do not close the lmdb.Env in the sync tests, because for some reason it
causes random SEGFAULTs on Linux (not on macOS).