[CI] Handle PRs from before check-spelling
Ideally check-spelling would happily handle PRs made from before
check-spelling was merged, but as of 0.0.21, it has a bug where it
doesn't do quite the right thing.

The `check-spelling/spell-check-this@prerelease` configuration
currently has a workaround where it includes a :steam_locomotive:
to suggest to users to rebase to a point where the configuration is
live.

I believe I've properly fixed the underlying bug in
`check-spelling/checkout-merge@prerelease`, but I need to test it for
a while before I release it, and all the features that are currently
in `check-spelling/check-spelling@prerelease` also need testing before
it'd be available in a `check-spelling/check-spelling@v0.0.22`.

In the interim, there's a fancier trick which should result in users
getting the desired behavior without requiring any additional updates
from check-spelling.

Instead of pointing `spell_check_this` to `check-spelling/spell-check-this@prerelease`
which is a default from `.github/workflows/spelling.yml` in
`check-spelling/spell-check-this@prerelease`,
we point to this repository's default branch
(`powerdns/lightningstream@main`).

The `spell_check_this` field is used by check-spelling whenever it
encounters a checkout without *any* configuration in the configuration
directory (by default, that's `.github/actions/spelling`).
Anyone who makes a branch from this repository after check-spelling
was added (0365b9d) should trigger this code path (unless they choose to
delete all of the configuration, which seems unlikely).
Thus, the only people who should be impacted are those who have created
a branch from before that commit and then try to create a PR to some
branch that includes check-spelling (and this commit).
In all likelihood, they will be creating their PR against the
`powerdns/lightningstream` repository and against the `main` branch.
As it'd be pretty weird to be creating a PR against that repository and
some other branch that also has check-spelling (instead of starting from
a more recent commit in case someone wanted to make a backport).

For more details about this workaround, see:
https://github.com/check-spelling/check-spelling/wiki/Onboarding#v0021